+++ src/main/java/it/polimi/se2018/model/objectives/privateobjectives/ShadesOfRedObjective.java
@@ -27,7 +27,7 @@ public class ShadesOfRedObjective extends PrivateObjective {
 
     @Override
     public int evalPoints(Player player){
-        return (int) StreamSupport.stream(player.getMap().spliterator(),false)
+        return StreamSupport.stream(player.getMap().spliterator(),false)
                 .map(Square::getDie)
                 .filter(Objects::nonNull)
                 .filter(die -> die.getColor()==color)
